Add developer tooling and fix static analysis defect findings
- Add justfile providing recipes for build, test, cppcheck, and clang-tidy
- Add .clang-tidy configured for C++11 and project include paths
- Add AGENTS.md guidelines for AI coding assistants and contributor instructions in BUILD.md
- Fix array bounds out-of-range indexing in symmetries_0.cpp, symmetries_1.cpp, symmetries_2.cpp
- Add [[noreturn]] attribute to bt_te() to aid static analyzer path pruning
- Fix potential null pointer dereferences and unsafe realloc patterns in Fl_Table.h and stlexportsolution.cpp
- Fix negative bit-shift and depth check logic in xml.cpp
- Fix signed/unsigned loop index in assembly.cpp and explicit string comparisons in assembler_0.cpp
- Add self-assignment check to vector3::operator= and unhide Fl_Double_Window::hide in multilinewindow.h

+++ src/lib/voxel_3.cpp
@@ -621,9 +621,9 @@ void voxel_3_c::getConnectionFace(int x, int y, int z, int n, double bevel, doub
 }
 
 void voxel_3_c::calculateSize(float * x, float * y, float * z) const {
-  *x = 2*((getX()+4)/5);
-  *y = 2*((getY()+4)/5);
-  *z = 2*((getZ()+4)/5);
+  *x = (float)(2 * ((getX() + 4) / 5));
+  *y = (float)(2 * ((getY() + 4) / 5));
+  *z = (float)(2 * ((getZ() + 4) / 5));
 }
 
 void voxel_3_c::recalcSpaceCoordinates(float * x, float * y, float * z) const {
